chrone: update project structure
