--- ui-react/pages/_document.tsx
@@ -1,23 +1,0 @@
-import { NextPageContext } from 'next';
-import Document, { Html, Head, Main, NextScript } from 'next/document'
-
-class MyDocument extends Document {
-  static async getInitialProps(ctx: NextPageContext) {
-    const initialProps = await Document.getInitialProps(ctx as any)
-    return { ...initialProps }
-  }
-
-  render() {
-    return (
-      <Html className="light">
-        <Head />
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    )
-  }
-}
-
-export default MyDocument
