+++ ui-react/src/components/layout/index.tsx
@@ -1,7 +1,6 @@
-import React from 'react';
 import Head from 'next/head';
-import Header from './Header';
-import styles from './Layout.module.css';
+import Header from './header';
+import styles from './layout.module.css';
 
 interface LayoutProps {
   children: React.ReactNode;
