+++ ui-react/src/components/pagination/index.tsx
@@ -1,5 +1,4 @@
-import React from "react";
-import styles from "./Pagination.module.css";
+import styles from "./pagination.module.css";
 
 const Pagination = () => {
   return (
