+++ ui-react/src/components/search-bar/search-bar.module.css
