+++ ui-react/src/pages/_app.tsx
@@ -1,9 +1,11 @@
 import '@moai/core/dist/bundle.css';
 import '@moai/core/dist/font/remote.css';
-import '../styles/globals.css';
+import 'styles/globals.css';
+
 import type { AppProps } from 'next/app';
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
+  return <Component {...pageProps} />;
 }
-export default MyApp
+
+export default MyApp;
