+++ ui-react/src/pages/index.tsx
@@ -1,5 +1,5 @@
 import { Button } from '@moai/core';
-import styles from '../styles/Home.module.css';
+import styles from 'styles/home.module.css';
 
 export default function Home() {
   return (
